im setting up a simple server
